Llamaindex/Basic_Retrieval_using_llama/requirements.txt
llama-index
python-dotenv
html2text 